+++ src/api/serializers.py
@@ -1,0 +1,17 @@
+from djoser.serializers import UserSerializer
+from rest_framework import serializers
+
+from disputes.models import Dispute
+from users.models import CustomUser
+
+
+class CustomUserSerializer(UserSerializer):
+    class Meta:
+        model = CustomUser
+        fields = '__all__'
+
+
+class DisputeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Dispute
+        fields = '__all__'
