--- src/users/serializers.py
@@ -1,9 +1,0 @@
-from djoser.serializers import UserSerializer
-
-from users.models import CustomUser
-
-
-class CustomUserSerializer(UserSerializer):
-    class Meta:
-        model = CustomUser
-        fields = '__all__'
